Work done during my time at The Alan Turing Insititute
During my 8 week internship at The Alan Turing Institute, I got the chance to play around with sqlsynthgen (https://github.com/alan-turing-institute/sqlsynthgen) which is a piece of software that can be used to create synthetic data. During my time, I began working on a config.yaml file and a generator.py file that could be used to create a synthetic version of an OMOP database (using sqlsynthgen). I also started work on a notebook that can be used to analyse most standard OMOP databases. The files include a requirements.txt file that have the requirements for the notebook to run.